requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
apscheduler==3.10.4
pytz==2024.1
groq==0.11.0
pyzbar==0.1.9
pillow==10.4.0
aiohttp==3.10.11
